designing quiz

+++ quiz_list.php
@@ -4,21 +4,44 @@
     <meta charset="utf-8">
     <title>Quiz list</title>
     <link rel="stylesheet" type="text/css" href="css/createQuiz.css">
+    <style media="screen">
+    table {
+      background-color: #fff;
+      font-family: arial, sans-serif;
+      border-collapse: collapse;
+      width: 85%;
+      margin-left: auto;
+      margin-right: auto;
+      font-size: 0.8em;
+      margin-bottom: 5em;
+
+    }
+    tr, th {
+      border: 1px solid #dddddd;
+      text-align: center;
+      align-items: center;
+      padding: 8px;
+    }
+    </style>
   </head>
   <body>
 
-    <?php include_once "navbaradmin.php"; ?>
+    <?php include_once("navbaradmin.php"); ?>
+
 
     <br><br><br><br>
-      <table height='50px' width='85%' class="list" >
 
+    <a href='quiz_title.php' class='addquiz'>Add Quiz</a><br><br>
+      <table height='50px' width='85%'>
         <tr>
-          <th colspan='2'><h2><label>title: </label></h2>
+          <th colspan='2'><h2><label><b>title: </b></label></h2>
+          <th colspan='2'><h2><label>Description: </label></h2>
+          <th colspan='2'><h2><label>Categories: </label></h2>
+          <th colspan='2'><h2><label>Publication Date: </label></h2>
           <th colspan='2'><h2><label>Quiz Code: </label></h2>
           <th colspan='3'><h2><label>Action: </label></h2>
         </tr>
         <br>
-        <a href='quiz_title.php' class='addquiz'>Add Quiz</a>
 
     <?php
     //Step 1 Database Connectivity
@@ -36,23 +59,44 @@
       while ($fetchTitle = mysqli_fetch_assoc($execQuery)) {
       $title = $fetchTitle["title"];
       $code = $fetchTitle['quiz_code'];
+      $Desc = $fetchTitle['description'];
+      $Cat = $fetchTitle['categories'];
+      $Pub = $fetchTitle['publish'];
 
-      echo "
-          <tr>
-            <br><br>
-            <th colspan='2'><h2>$title</h2></th>
-            <th colspan='2'><h2>$code</h2></th>
-            <th colspan='3'>
-              <a href='editTitle.php?quiz_code=$code' class='action'>edit title</a>
-              <a href='questions.php?quiz_code=$code' class='action'>edit/add questions</a>
-              <a href='deleteQuiz.php?quiz_code=$code' class='action'>delete</a>
-            </th>
-          </tr>
 
-      ";
-    }
-  }
-    }
+      if(strlen($title) >= 20){
+            $title = substr($title,0,20) . "...";
+
+            if(strlen($Desc) >= 10){
+                  $Desc = substr($Desc,0,10) . "...";
+
+                  $newDate = date("m-d-Y", strtotime($Pub));
+
+                  echo "
+                      <tr>
+                        <br><br>
+                        <th colspan='2'>
+                          <div>
+                            <h2>$title</h2></th>
+                          </div>
+                        <th colspan='2'><h2>$Desc</h2></th>
+                        <th colspan='2'><h2>$Cat</h2></th>
+                        <th colspan='2'><h2>$newDate</h2></th>
+                        <th colspan='2'><h2>$code</h2></th>
+
+                        <th colspan='3'>
+                          <a href='editTitle.php?quiz_code=$code' class='action'>edit title</a>
+                          <a href='questions.php?quiz_code=$code' class='action'>edit/add questions</a>
+                          <a href='deleteQuiz.php?quiz_code=$code' class='action'>delete</a>
+                        </th>
+                      </tr>
+
+                  ";
+                }
+              }
+            }
+          }
+        }
      ?>
 
    </table>
